Stop too many bytes for activequeues.
We were saying calloc(N,N*sizeof(struct event_list*)) when we should have
been saying calloc(N,sizeof(struct event_list*)).  This wasted N*(N-1) words
of memory, where N was the number of priorities.  This wouldn't be a big deal
for any sane number of priorities, but it's a bug nonetheless.

git-svn-id: svn://svn.transmissionbt.com/libevent/branches/patches-1.4/libevent@1526 48a2d78c-3def-41fb-ae5e-7081b2a5aef7