Add support for Git to update-version-h.sh
Since it's not easily possible to generate linear version numbers in git,
simply use zero. Also append '-dirty' to the source version string if the
working tree contains uncommitted changes.
In case there is neither a svn nor git repository and we try to guess the
revision number, append '(guessed)' to the version string so it's apparent
that the version might not be accurate.

Signed-off-by: Tomas Carnecky <tom@dbservice.com>