Use a port range when binding the local port
If the bind() fails try the next higher port. Repeat for 50 ports before
definitely failing. This fixes two issues: first, binding to the same port
as the incoming port doesn't work at all under Mac OS X. When bind-local-port
is enabled there the bind() will always fail the first time and then use
the next higher port. Second, this allows us to connect to the same peer
multiple times (in case we have found a peer which has two or more torrents
as we have).